Correct Groovy limitation over inner classes for previous builder pattern code
